+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added scm-manager plugin
+- Volume for /tmp directory. This volume is not included in backups!
+
+### Changed
+- Upgrade to Jenkins 2.222.1; #52
+- Upgrade java base image to 8u242-1
 
 ## [v2.190.3-3] - 2020-02-26
 ### Added